+++ rtlib/setting.py
@@ -12,7 +12,6 @@ import discord
 
 from pytz import utc
 
-from aiohttp import ClientSession
 from ujson import dumps
 
 from rtlib.rt_module.src.setting import CommandData, CommandRunData
@@ -116,7 +115,7 @@ class SettingManager(commands.Cog):
 
     def session(self):
         "`aiohttp.ClientSession`の準備をする。"
-        return ClientSession(loop=self.bot.loop, json_serialize=dumps)
+        return self.bot.session
 
     def _get_default(self, default: object) -> object:
         # 渡されたものがスラッシュのオプションならそれに設定されているデフォルトを返す。