+++ util/markdowns.py
@@ -26,7 +26,7 @@ def separate(text: str, character: str = "\n") -> Tuple[str, str]:
     return text[:(i:=text.find(character))], text[i+1:]
 
 
-def create_embed_from(markdown: str, **kwargs) -> Embed:
+def create_embed(markdown: str, **kwargs) -> Embed:
     """渡されたマークダウンの文字列をタイトルと説明とフィールドが設定されている`discord.Embed`に変換します。  
     見出しは三段階設定することができ、一段でタイトルで二段でフィールドそして三段で`**#** ...`のようになります。  
     もしフィールドの`inline`を`False`にしたい場合は`## !`のようにしてください。